renamed jwt and local strats
