+++ passport/jwtStrategy.js
